+++ backend/src/db/errors/DBDeleteErrors.ts
@@ -1,4 +1,4 @@
-import { BaseError } from '@Infra/Result';
+import { BaseError } from '@infra/BaseError';
 
 export class DBDeleteError extends BaseError {
   constructor(public tableName: string) {
