+++ backend/src/db/errors/DBSelectErrors.ts
@@ -1,4 +1,4 @@
-import { BaseError } from '@Infra/Result';
+import { BaseError } from '@infra/BaseError';
 
 export class DBSelectError extends BaseError {
   constructor(public tableName: string) {
