+++ backend/src/game/PokerTable.test.ts
@@ -1,7 +1,7 @@
 // Internal
-import { createPokerTable } from '@Tests/helpers/createPokerTable';
-import { createPokerTableWithPlayers } from '@Tests/helpers/createPokerTableWithPlayers';
-import { mockSendEventToRoomSuccess } from '@Tests/mocks/roomMocks';
+import { createPokerTable } from '@tests/helpers/createPokerTable';
+import { createPokerTableWithPlayers } from '@tests/helpers/createPokerTableWithPlayers';
+import { mockSendEventToRoomSuccess } from '@tests/mocks/roomMocks';
 
 describe('PokerTable', () => {
   describe('sitAtTable', () => {
