+++ backend/src/game/PokerTable.ts
@@ -1,17 +1,17 @@
 // Internal
-import { Result, ResultSuccess, BaseError } from '@Infra/Result';
+import { Result, ResultSuccess } from '@infra/Result';
 import { Seat } from './Seat';
-import { PlayerAlreadySeatedError, SeatTakenError, PlayerNotFoundAtTableError } from '../shared/api/PokerTables/errors';
+
+import {
+  PlayerAlreadySeatedError,
+  PlayerNotFoundAtTableError,
+  SeatNotFoundError,
+  SeatTakenError,
+} from '../handlers/poker-tables/errors';
 
 /* 
   PokerTable is responsible for managing a single poker table. It will be responsible for managing the game state (which is a class), players at a table..
 */
-
-class SeatNotFoundError extends BaseError {
-  constructor() {
-    super('SEAT_NOT_FOUND', 'Seat not found');
-  }
-}
 
 export class PokerTable {
   private tableName: string;
