+++ backend/src/handlers/helpers/mapBaseErrorToAPIError.ts
@@ -1,4 +1,4 @@
-import { IBaseError } from '@Infra/Result';
+import { IBaseError } from '@infra/BaseError';
 import { IAPIError } from '../../shared/api/BaseOutput';
 
 export function mapBaseErrorToAPIError(baseError: IBaseError): IAPIError {
