+++ backend/src/handlers/poker-tables/PokerTablesLeaveHandler.test.ts
@@ -4,9 +4,9 @@ import request from 'supertest';
 // Internal
 import { httpServer } from '../../index';
 import { GameLobbyService } from '../../game-lobby-service';
-import { shutDownServer } from '@Tests/helpers/shutDownServer';
+import { shutDownServer } from '@tests/helpers/shutDownServer';
 import { Rooms } from '../../sockets/Rooms';
-import { ResultSuccess, ResultError } from '@Infra/Result';
+import { ResultSuccess, ResultError } from '@infra/Result';
 import { RoomNotFoundError } from '../../sockets/errors/RoomErrors';
 
 describe('tables.leave', () => {
