+++ backend/src/handlers/poker-tables/errors/PokerTableNameTakenError.ts
@@ -1,4 +1,4 @@
-import { BaseError } from '@Infra/Result';
+import { BaseError } from '@infra/BaseError';
 
 export class PokerTableNameTakenError extends BaseError {
   constructor() {
