+++ backend/src/handlers/poker-tables/errors/SeatNotFoundError.ts
@@ -1,0 +1,7 @@
+import { BaseError } from '@infra/BaseError';
+
+export class SeatNotFoundError extends BaseError {
+  constructor() {
+    super('SEAT_NOT_FOUND', 'Seat not found');
+  }
+}
