+++ backend/src/handlers/poker-tables/errors/index.ts
@@ -1,0 +1,17 @@
+import { PlayerAlreadySeatedError } from './PlayerAlreadySeatedError';
+import { PlayerNotFoundAtTableError } from './PlayerNotFoundAtTableError';
+
+import { PokerTableDoesNotExistError } from './PokerTableDoesNotExistError';
+import { PokerTableNameTakenError } from './PokerTableNameTakenError';
+
+import { SeatNotFoundError } from './SeatNotFoundError';
+import { SeatTakenError } from './SeatTakenError';
+
+export {
+  PlayerAlreadySeatedError,
+  PlayerNotFoundAtTableError,
+  PokerTableDoesNotExistError,
+  PokerTableNameTakenError,
+  SeatNotFoundError,
+  SeatTakenError,
+};
