+++ backend/src/handlers/users/errors/index.ts
@@ -1,0 +1,5 @@
+import { UserNameTakenError } from './UserNameTakenError';
+
+import { UsersCreateError } from './UsersCreateError';
+
+export { UserNameTakenError, UsersCreateError };
