+++ backend/src/shared/api/ApiMethodMap.ts
@@ -3,9 +3,9 @@ import type { NextFunction, Request, Response } from 'express';
 import type { BaseOutput } from './BaseOutput';
 
 // Handler types
-import type { PokerTableJoinPayload, PokerTableJoinOutput } from './PokerTables/types/PokerTableJoin';
-import type { PokerTableLeavePayload, PokerTableLeaveOutput } from './PokerTables/types/PokerTableLeave';
-import type { UsersCreatePayload, UsersCreateOutput } from './Users/types/UsersCreate';
+import type { PokerTableJoinPayload, PokerTableJoinOutput } from './poker-tables/types/PokerTableJoin';
+import type { PokerTableLeavePayload, PokerTableLeaveOutput } from './poker-tables/types/PokerTableLeave';
+import type { UsersCreatePayload, UsersCreateOutput } from './users/types/UsersCreate';
 
 export interface ApiHandler {
   runHandler(req: Request<any>, res: Response<BaseOutput>, next: NextFunction): any;
