--- backend/src/shared/api/PokerTables/errors/index.ts
@@ -1,6 +1,0 @@
-import { PlayerAlreadySeatedError } from './PlayerAlreadySeatedError';
-import { SeatTakenError } from './SeatTakenError';
-import { PokerTableDoesNotExistError } from './PokerTableDoesNotExistError';
-import { PlayerNotFoundAtTableError } from './PlayerNotFoundAtTableError';
-
-export { PlayerAlreadySeatedError, SeatTakenError, PokerTableDoesNotExistError, PlayerNotFoundAtTableError };
