--- backend/src/shared/api/Users/errors/index.ts
@@ -1,4 +1,0 @@
-import { UsersCreateError } from './UsersCreateError';
-import { UserNameTakenError } from './UserNameTakenError';
-
-export { UsersCreateError, UserNameTakenError };
