+++ backend/src/sockets/errors/RoomErrors.ts
@@ -1,4 +1,4 @@
-import { BaseError } from '@Infra/Result';
+import { BaseError } from '@infra/BaseError';
 
 export class RoomAlreadyExistsError extends BaseError {
   constructor() {
