+++ backend/src/sockets/errors/SocketErrors.ts
@@ -1,4 +1,4 @@
-import { BaseError } from '@Infra/Result';
+import { BaseError } from '@infra/BaseError';
 
 export class SocketNotFoundError extends BaseError {
   constructor() {
